Add some performance metrics and some early-exit optimizations
If Atom is in dev mode performance metrics of how long the total lint took, as well as how lpong JSCS itself took are output to the console. Also fixes some early-exit scenarios such as when a file was ignored by the configuration. Previously JSCS was fully initialized before checking if the file was ignored.