Update CI config (#296)
* Update CI config

Ensure's Node.js v6 is installed on the system, and moves to a script that uses the system Node.js for linting. This should be switched back to the atom/ci scripts when a stable Atom release includes an `apm` with a usable `node` version.

* Add `temp` to required devDependencies

Looks like this wasn't added before, I'm not sure how the issue didn't come up, but it has now...

Also tells `temp` to track the files it creates.